💣 Sleep/Wake Button

+++ src/deep_sleep.h
@@ -1,3 +1,3 @@
 void wake_up_and_react();
 void send_sleeping();
-long getSecondsTillWakeUp();
+int getSecondsTillWakeUp();
